Adds getSessionAttributes to updateSessionOperation interface and move logic for getting session attributes into biometricTokenIssued
