+++ backend-api/src/functions/common/session/updateOperations/BiometricTokenIssued/BiometricTokenIssued.ts
@@ -2,9 +2,12 @@ import { UpdateSessionOperation } from "../UpdateSessionOperation";
 import { DocumentType } from "../../../../types/document";
 import { SessionState } from "../../session";
 import {
+  AttributeValue,
   ReturnValue,
   ReturnValuesOnConditionCheckFailure,
 } from "@aws-sdk/client-dynamodb";
+import { BaseSessionAttributes } from "../../../../adapters/dynamoDbAdapter";
+import { NativeAttributeValue, unmarshall } from "@aws-sdk/util-dynamodb";
 
 export class BiometricTokenIssued implements UpdateSessionOperation {
   constructor(
@@ -32,7 +35,45 @@ export class BiometricTokenIssued implements UpdateSessionOperation {
   getDynamoDbReturnValues() {
     return ReturnValue.ALL_NEW;
   }
+
   getDynamoDbReturnValuesOnConditionCheckFailure() {
     return ReturnValuesOnConditionCheckFailure.ALL_OLD;
   }
+
+  getSessionAttributes(item: Record<string, NativeAttributeValue>) {
+    return getBaseSessionAttributes(item);
+  }
 }
+
+const getBaseSessionAttributes = (
+  item: Record<string, AttributeValue> | undefined,
+): BaseSessionAttributes | null => {
+  if (item == null) return null;
+
+  const sessionAttributes = unmarshall(item);
+  if (!isBaseSessionAttributes(sessionAttributes)) return null;
+
+  return sessionAttributes;
+};
+
+const isBaseSessionAttributes = (
+  item: Record<string, NativeAttributeValue>,
+): item is BaseSessionAttributes => {
+  if (typeof item.clientId !== "string") return false;
+  if (typeof item.govukSigninJourneyId !== "string") return false;
+  if (typeof item.createdAt !== "number") return false;
+  if (typeof item.issuer !== "string") return false;
+  if (typeof item.sessionId !== "string") return false;
+  if (typeof item.sessionState !== "string") return false;
+  if (typeof item.clientState !== "string") return false;
+  if (typeof item.subjectIdentifier !== "string") return false;
+  if (typeof item.timeToLive !== "number") return false;
+  if (
+    "redirectUri" in item &&
+    item.redirectUri !== undefined &&
+    typeof item.redirectUri !== "string"
+  ) {
+    return false;
+  }
+  return true;
+};
